Cleaned up the naming and usage
